criando header da pokedex e detalhes
