+++ pokedex/src/pages/homePage.js
@@ -3,12 +3,14 @@ import {goToDetalhesPage} from "../routes/coordinator"
 import {goToPokedexPage} from "../routes/coordinator"
 import { useHistory } from "react-router";
 import styled from 'styled-components'
+import { HeaderPokedex } from "../componetes/headerPokedex";
 
 
 export const HomePage = () => {
     const history = useHistory();
     return(
     <div>
+        <HeaderPokedex />
         <p>Home</p>
         <CardPoke>
             <Img src={"https://img.pokemondb.net/artwork/charmeleon.jpg"} />
