+++ pokedex/src/pages/pokedexPage.js
@@ -2,11 +2,13 @@ import React from "react"
 import {goToDetalhesPage} from "../routes/coordinator"
 import { useHistory } from "react-router";
 import styled from 'styled-components'
+import { HeaderPokedex } from "../componetes/headerPokedex";
 
 export const PokedexPage = () => {
     const history = useHistory();
     return(
     <div>
+        <HeaderPokedex/>
         <p>Pokedex</p>
         <CardPoke>
             <Img src={"https://img.pokemondb.net/artwork/charmeleon.jpg"} />
